Split `MessageValidationCode` into data and operation
